now it works even when sending more than 1 message

+++ main.py
@@ -6,22 +6,25 @@ POST_ENDPOINT = "https://ShortChat.elio27.repl.co/api/chat/post"
 READ_ENDPOINT = "https://ShortChat.elio27.repl.co/api/chat"
 
 if not os.path.exists("token.txt"):
-	name = input("Welcome to the ShortChat Python script!\nChoose a username: ")
-	token_url = TOKEN_ENDPOINT + name
-	token = requests.get(token_url).text
+  
+  name = input("Welcome to the ShortChat Python script!\nChoose a username: ")
+  token_url = TOKEN_ENDPOINT + name
+  token = requests.get(token_url).text
+  
+  if token == "This username is already taken":
+    print("This username is already taken")
+  else:
+    with open("token.txt", "w") as f:
+        f.write(token)
 
-	with open("token.txt", "w") as f:
-		f.write(token)
+message = ""
+while message != "exit":
+  with open("token.txt") as token:
+      print(requests.get(READ_ENDPOINT).text.strip() + "\n")
 
-with open("token.txt") as token:
-	message = ""
+      message = input("Send a message or refresh by pressing enter.\n>>> ")
 
-	while message != "exit":
-		print(requests.get(READ_ENDPOINT).text.strip() + "\n")
-
-		message = input("Send a message or refresh by pressing enter.\n>>> ")
-
-		if message:
-			requests.post(POST_ENDPOINT, data = {"uuid": token.read(), "message": message})
-		
-		os.system("cls" if os.name == "nt" else "clear")
+      if message:
+          requests.post(POST_ENDPOINT, data = {"uuid": token.read(), "message": message})
+      
+      os.system("cls" if os.name == "nt" else "clear")
